requirements.txt
streamlit==0.83.0
pillow==8.2.0
numpy==1.18.5
pandas==1.2.4
matplotlib==3.4.2
plotly-express==0.4.1
seaborn==0.11.1